+++ CHIP_IO/LRADC.py
@@ -42,9 +42,12 @@ CURRENT_SAMPLE_RATE_FILE = "/in_voltage_sampling_frequency"
 RAW_VOLTAGE_CHAN0_FILE = "/in_voltage0_raw"
 RAW_VOLTAGE_CHAN1_FILE = "/in_voltage1_raw"
 
-def enable_debug():
+def toggle_debug():
     global DEBUG
-    DEBUG = True
+    if DEBUG:
+        DEBUG = False
+    else:
+        DEBUG = True
 
 def setup(rate=250):
     # First we determine if the device exists
@@ -75,7 +78,7 @@ def get_scale_factor():
 
     # Debug
     if DEBUG:
-        print("Current LRADC Scaling Factor: {0}".format(SCALE_FACTOR))
+        print("lradc.get_scale_factor: {0}".format(SCALE_FACTOR))
 
     return SCALE_FACTOR
 
@@ -100,7 +103,7 @@ def get_allowable_sample_rates():
 
     # Debug
     if DEBUG:
-        print("Allowable Sampling Rates:")
+        print("lradc.get_allowable_sample_rates:")
         for rate in SAMPLE_RATES:
             print("{0}".format(rate))
 
@@ -122,7 +125,7 @@ def set_sample_rate(rate):
 
     # Debug
     if DEBUG:
-        print("Setting Sample Rate to: {0}".format(rate))
+        print("lradc.set_sample_rate: {0}".format(rate))
 
     # Write the rate
     f = open(LRADC_BASE_DEVICE_FILE+CURRENT_SAMPLE_RATE_FILE,"w")
@@ -153,7 +156,7 @@ def get_sample_rate():
 
     # Debug
     if DEBUG:
-        print("Current Samplig Rate: {0}".format(dat))
+        print("lradc.get_sample_rate: {0}".format(dat))
 
     return dat
 
@@ -171,7 +174,7 @@ def get_chan0_raw():
 
     # Debug
     if DEBUG:
-        print("CHAN0 RAW: {0}".format(dat))
+        print("lradc.get_chan0_raw: {0}".format(dat))
 
     return dat
 
@@ -189,7 +192,7 @@ def get_chan1_raw():
 
     # Debug
     if DEBUG:
-        print("CHAN1 RAW: {0}".format(dat))
+        print("lradc.get_chan1_raw: {0}".format(dat))
 
     return dat
 
@@ -205,7 +208,7 @@ def get_chan0():
 
     # Debug
     if DEBUG:
-        print("CHAN0: {0}".format(dat))
+        print("lradc.get_chan0: {0}".format(dat))
 
     return dat
 
@@ -221,7 +224,7 @@ def get_chan1():
 
     # Debug
     if DEBUG:
-        print("CHAN1: {0}".format(dat))
+        print("lradc.get_chan1: {0}".format(dat))
 
     return dat