+++ source/c_pwm.c
@@ -33,6 +33,7 @@ SOFTWARE.
 #include <stdlib.h>
 #include <sys/types.h>
 #include <string.h>
+#include <errno.h>
 #include <fcntl.h>
 #include <unistd.h>
 #include "c_pwm.h"
@@ -47,9 +48,6 @@ SOFTWARE.
 
 // Global variables
 int pwm_initialized = 0;
-int DEBUG = 0;
-//int ENABLE = 1;
-//int DISABLE = 0;
 
 // pwm devices (future chip pro use)
 struct pwm_dev
@@ -105,6 +103,7 @@ struct pwm_exp *lookup_exported_pwm(const char *key)
 
 int initialize_pwm(void)
 {
+    int e_no;
     if (!pwm_initialized) {
         int fd, len;
         char str_gpio[80];
@@ -113,26 +112,40 @@ int initialize_pwm(void)
         int gpio = 0;
 
         if (DEBUG)
-            printf(" ** EXPORTING PWM0 **\n");
+            printf(" ** initialize_pwm **\n");
         if ((fd = open("/sys/class/pwm/pwmchip0/export", O_WRONLY)) < 0)
         {
+            char err[256];
+            snprintf(err, sizeof(err), "initialize_pwm: could not open export file");
+            add_error_msg(err);
             return -1;
         }
         len = snprintf(str_gpio, sizeof(str_gpio), "%d", gpio); BUF2SMALL(str_gpio);
-        ssize_t s = write(fd, str_gpio, len);  ASSRT(s == len);
-        if (DEBUG)
-            printf(" ** IN initialize_pwm: s = %d, len = %d\n", s, len);
+        ssize_t s = write(fd, str_gpio, len); e_no = errno;
         close(fd);
-
+        
+        if (s != len) {
+            char err[256];
+            snprintf(err, sizeof(err), "initialize_pwm: could not export pwm (%s)", strerror(e_no));
+            add_error_msg(err);
+            return -1;
+        }
+        
+        if (DEBUG)
+            printf(" ** initialize_pwm: export pin: s = %d, len = %d\n", s, len);
+        
         pwm_initialized = 1;
         return 1;
+    } else {
+        if (DEBUG)
+            printf(" ** initialize_pwm: pwm is already initialized\n");
     }
 
     return 0;
 }
 
 int pwm_set_frequency(const char *key, float freq) {
-    int len;
+    int len, e_no;
     int rtnval = -1;
     char buffer[80];
     unsigned long period_ns;
@@ -150,18 +163,20 @@ int pwm_set_frequency(const char *key, float freq) {
     period_ns = (unsigned long)(1e9 / freq);
 
     if (pwm->enable) {
-        if (DEBUG)
-            printf(" ** IN pwm_set_frequency: pwm_initialized = %d\n", pwm_initialized);
         if (period_ns != pwm->period_ns) {
             pwm->period_ns = period_ns;
 
             len = snprintf(buffer, sizeof(buffer), "%lu", period_ns); BUF2SMALL(buffer);
-            if (DEBUG)
+            ssize_t s = write(pwm->period_fd, buffer, len);  e_no = errno;
+            if (DEBUG) {
+                printf(" ** pwm_set_frequency: pwm_initialized = %d\n", pwm_initialized);
                 printf(" ** pwm_set_frequency: buffer: %s\n", buffer);
-            ssize_t s = write(pwm->period_fd, buffer, len);  //ASSRT(s == len);
-            if (DEBUG)
-                printf(" ** IN pwm_set_frequency: s = %d, len = %d\n", s, len);
+                printf(" ** pwm_set_frequency: s = %d, len = %d\n", s, len);
+            }
             if (s != len) {
+                char err[256];
+                snprintf(err, sizeof(err), "pwm_set_frequency: could not change frequency of pwm (%s)", strerror(e_no));
+                add_error_msg(err);
                 rtnval = -1;
             } else {
                 rtnval = 1;
@@ -176,8 +191,71 @@ int pwm_set_frequency(const char *key, float freq) {
     return rtnval;
 }
 
+int pwm_set_period_ns(const char *key, unsigned long period_ns) {
+    int len, e_no;
+    int rtnval = -1;
+    char buffer[80];
+    struct pwm_exp *pwm;
+
+    //TODO: ADD CHECK FOR period_ns
+
+    pwm = lookup_exported_pwm(key);
+
+    if (pwm == NULL) {
+        return rtnval;
+    }
+    
+    if (pwm->enable) {
+        if (period_ns != pwm->period_ns) {
+            pwm->period_ns = period_ns;
+
+            len = snprintf(buffer, sizeof(buffer), "%lu", period_ns); BUF2SMALL(buffer);
+            ssize_t s = write(pwm->period_fd, buffer, len);  e_no = errno;
+            if (DEBUG) {
+                printf(" ** pwm_set_period_ns: pwm_initialized = %d\n", pwm_initialized);
+                printf(" ** pwm_set_period_ns: buffer: %s\n", buffer);
+                printf(" ** pwm_set_period_ns: s = %d, len = %d\n", s, len);
+            }
+            if (s != len) {
+                char err[256];
+                snprintf(err, sizeof(err), "pwm_set_period_ns: could not change period of pwm (%s)", strerror(e_no));
+                add_error_msg(err);
+                rtnval = -1;
+            } else {
+                rtnval = 1;
+            }
+        } else {
+            rtnval = 0;
+        }
+    } else {
+        rtnval = 0;
+    }
+
+    return rtnval;
+}
+
+int pwm_get_period_ns(const char *key, unsigned long *period_ns) {
+    int rtnval = -1;
+    struct pwm_exp *pwm;
+    
+    pwm = lookup_exported_pwm(key);
+
+    if (pwm == NULL) {
+        return rtnval;
+    }
+    
+    if (DEBUG)
+        printf(" ** pwm_get_period_ns: %lu **\n",pwm->period_ns);
+    
+    // Set period_ns to what we have in the struct
+    *period_ns = pwm->period_ns;
+    
+    rtnval = 0;
+    return rtnval;
+}
+
 int pwm_set_polarity(const char *key, int polarity) {
-    int len;
+    int len, e_no;
     int rtnval = -1;
     char buffer[80];
     struct pwm_exp *pwm;
@@ -193,8 +271,6 @@ int pwm_set_polarity(const char *key, int polarity) {
     }
 
     if (pwm->enable) {
-        if (DEBUG)
-            printf(" ** IN pwm_set_polarity: pwm_initialized = %d\n", pwm_initialized);
         if (polarity == 0) {
             len = snprintf(buffer, sizeof(buffer), "%s", "normal"); BUF2SMALL(buffer);
         }
@@ -202,12 +278,16 @@ int pwm_set_polarity(const char *key, int polarity) {
         {
             len = snprintf(buffer, sizeof(buffer), "%s", "inverted"); BUF2SMALL(buffer);
         }
-        if (DEBUG)
-            printf(" ** pwm_set_poliarity: buffer: %s\n", buffer);
-        ssize_t s = write(pwm->polarity_fd, buffer, len);  //ASSRT(s == len);
-        if (DEBUG)
-            printf(" ** IN pwm_set_polarity: s = %d, len = %d\n", s, len);
+        ssize_t s = write(pwm->polarity_fd, buffer, len);  e_no = errno;
+        if (DEBUG) {
+            printf(" ** pwm_set_polarity: pwm_initialized = %d\n", pwm_initialized);
+            printf(" ** pwm_set_polarity: buffer: %s\n", buffer);
+            printf(" ** pwm_set_polarity: s = %d, len = %d\n", s, len);
+        }
         if (s != len) {
+            char err[256];
+            snprintf(err, sizeof(err), "pwm_set_polarity: could not change polarity of pwm (%s)", strerror(e_no));
+            add_error_msg(err);
             rtnval = -1;
         } else {
             rtnval = 1;
@@ -219,7 +299,7 @@ int pwm_set_polarity(const char *key, int polarity) {
 }
 
 int pwm_set_duty_cycle(const char *key, float duty) {
-    int len;
+    int len, e_no;
     int rtnval = -1;
     char buffer[80];
     struct pwm_exp *pwm;
@@ -237,15 +317,17 @@ int pwm_set_duty_cycle(const char *key, float duty) {
     pwm->duty = (unsigned long)(pwm->period_ns * (duty / 100.0));
 
     if (pwm->enable) {
-        if (DEBUG)
-            printf(" ** IN pwm_set_duty_cycle: pwm_initialized = %d\n", pwm_initialized);
         len = snprintf(buffer, sizeof(buffer), "%lu", pwm->duty); BUF2SMALL(buffer);
-        if (DEBUG)
+        ssize_t s = write(pwm->duty_fd, buffer, len);  e_no = errno;
+        if (DEBUG) {
+            printf(" ** pwm_set_duty_cycle: pwm_initialized = %d\n", pwm_initialized);
             printf(" ** pwm_set_duty_cycle: buffer: %s\n", buffer);
-        ssize_t s = write(pwm->duty_fd, buffer, len);  //ASSRT(s == len);
-        if (DEBUG)
-            printf(" ** IN pwm_set_duty_cycle: s = %d, len = %d\n", s, len);
+            printf(" ** pwm_set_duty_cycle: s = %d, len = %d\n", s, len);
+        }
         if (s != len) {
+            char err[256];
+            snprintf(err, sizeof(err), "pwm_set_duty_cycle: could not change duty cycle of pwm (%s)", strerror(e_no));
+            add_error_msg(err);
             rtnval = -1;
         } else {
             rtnval = 1;
@@ -257,16 +339,57 @@ int pwm_set_duty_cycle(const char *key, float duty) {
     return rtnval;
 }
 
+int pwm_set_pulse_width_ns(const char *key, unsigned long pulse_width_ns) {
+    int len, e_no;
+    int rtnval = -1;
+    char buffer[80];
+    struct pwm_exp *pwm;
+
+    pwm = lookup_exported_pwm(key);
+
+    if (pwm == NULL) {
+        return rtnval;
+    }
+
+    if (pulse_width_ns < 0 || pulse_width_ns > pwm->period_ns)
+        return rtnval;
+
+    pwm->duty = pulse_width_ns / pwm->period_ns;
+    
+    if (pwm->enable) {
+        len = snprintf(buffer, sizeof(buffer), "%lu", pwm->duty); BUF2SMALL(buffer);
+        ssize_t s = write(pwm->duty_fd, buffer, len);  e_no = errno;
+        if (DEBUG) {
+            printf(" ** pwm_set_pulse_width_ns: pwm_initialized = %d\n", pwm_initialized);
+            printf(" ** pwm_set_pulse_width_ns: buffer: %s\n", buffer);
+            printf(" ** pwm_set_pulse_width_ns: s = %d, len = %d\n", s, len);
+        }
+        if (s != len) {
+            char err[256];
+            snprintf(err, sizeof(err), "pwm_set_pulse_width_ns: could not change pulse width of pwm (%s)", strerror(e_no));
+            add_error_msg(err);
+            rtnval = -1;
+        } else {
+            rtnval = 1;
+        }
+    } else {
+        rtnval = 0;
+    }
+
+    return rtnval;
+       
+}
+
 int pwm_set_enable(const char *key, int enable)
 {
-    int len;
+    int len, e_no;
     int rtnval = -1;
     char buffer[80];
     struct pwm_exp *pwm;
 
     if (enable != 0 && enable != 1) {
         if (DEBUG)
-            printf(" ** INVALID ENABLE OPTION, NEEDS TO BE EITHER 0 OR 1! **\n");
+            printf(" ** pwm_set_enable: enable needs to be 0 or 1! **\n");
         return rtnval;
     }
 
@@ -274,26 +397,29 @@ int pwm_set_enable(const char *key, int enable)
 
     if (pwm == NULL) {
         if (DEBUG)
-            printf(" ** SOMETHING BAD HAPPEND IN pwm_set_enable, WE'RE EXITING WITH -1 NOW! **\n");
+            printf(" ** pwm_set_enable: pwm struct is null **\n");
         return rtnval;
     }
 
 
     len = snprintf(buffer, sizeof(buffer), "%d", enable); BUF2SMALL(buffer);
-    ssize_t s = write(pwm->enable_fd, buffer, len);  //ASSRT(s == len);
+    ssize_t s = write(pwm->enable_fd, buffer, len);  e_no = errno;
     if (DEBUG) {
-        printf(" ** IN pwm_set_enable: pwm_initialized = %d\n", pwm_initialized);
+        printf(" ** pwm_set_enable: pwm_initialized = %d\n", pwm_initialized);
         printf(" ** pwm_set_enable: buffer: %s\n", buffer);
-        printf(" ** IN pwm_set_enable: s = %d, len = %d\n", s, len);
+        printf(" ** pwm_set_enable: s = %d, len = %d\n", s, len);
     }
 
     if (s == len)
     {
         if (DEBUG)
-            printf(" ** SETTING pwm->enable to %d\n", enable);
+            printf(" ** pwm_set_enable: pwm->enable to %d\n", enable);
         pwm->enable = enable;
         rtnval = 0;
     } else {
+        char err[256];
+        snprintf(err, sizeof(err), "pwm_set_enable: could not enable/disable pwm (%s)", strerror(e_no));
+        add_error_msg(err);
         rtnval = -1;
     }
 
@@ -311,16 +437,16 @@ int pwm_start(const char *key, float duty, float freq, int polarity)
     struct pwm_exp *new_pwm, *pwm;
 
     if (DEBUG)
-        printf(" ** IN pwm_start: pwm_initialized = %d\n", pwm_initialized);
+        printf(" ** pwm_start: pwm_initialized = %d\n", pwm_initialized);
     if(!pwm_initialized) {
         initialize_pwm();
     } else {
         if (DEBUG)
-            printf(" ** ALREADY INITIALIZED, CALLING CLEANUP TO START FRESH **");
+            printf(" ** pwm_start: pwm already initialized, cleaning up **");
         pwm_cleanup();   
     }
     if (DEBUG)
-        printf(" ** IN pwm_start: pwm_initialized = %d\n", pwm_initialized);
+        printf(" ** pwm_start: pwm_initialized = %d\n", pwm_initialized);
 
     //setup the pwm base path, the chip only has one pwm
     snprintf(pwm_base_path, sizeof(pwm_base_path), "/sys/class/pwm/pwmchip0/pwm%d", 0); BUF2SMALL(pwm_base_path);
@@ -332,19 +458,26 @@ int pwm_start(const char *key, float duty, float freq, int polarity)
     snprintf(polarity_path, sizeof(polarity_path), "%s/polarity", pwm_base_path); BUF2SMALL(polarity_path);
 
     if (DEBUG) {
-        printf(" ** IN pwm_start: pwm_base_path: %s\n", pwm_base_path);
-        printf(" ** IN pwm_start: enable_path:   %s\n", enable_path);
-        printf(" ** IN pwm_start: period_path:   %s\n", period_path);
-        printf(" ** IN pwm_start: duty_path:     %s\n", duty_path);
-        printf(" **IN pwm_start: polarity_path: %s\n", polarity_path);
+        printf(" ** pwm_start: pwm_base_path: %s\n", pwm_base_path);
+        printf(" ** pwm_start: enable_path:   %s\n", enable_path);
+        printf(" ** pwm_start: period_path:   %s\n", period_path);
+        printf(" ** pwm_start: duty_path:     %s\n", duty_path);
+        printf(" ** pwm_start: polarity_path: %s\n", polarity_path);
     }
     
     //add period and duty fd to pwm list
-    if ((enable_fd = open(enable_path, O_WRONLY)) < 0)
+    if ((enable_fd = open(enable_path, O_WRONLY)) < 0) {
+        char err[256];
+        snprintf(err, sizeof(err), "pwm_start: could not open enable file");
+        add_error_msg(err);
         return -1;
+    }
 
     if ((period_fd = open(period_path, O_WRONLY)) < 0) {
         close(enable_fd);
+        char err[256];
+        snprintf(err, sizeof(err), "pwm_start: could not open period file");
+        add_error_msg(err);
         return -1;
     }
 
@@ -352,6 +485,9 @@ int pwm_start(const char *key, float duty, float freq, int polarity)
         //error, close already opened period_fd.
         close(enable_fd);
         close(period_fd);
+        char err[256];
+        snprintf(err, sizeof(err), "pwm_start: could not open duty cycle file");
+        add_error_msg(err);
         return -1;
     }
 
@@ -360,17 +496,23 @@ int pwm_start(const char *key, float duty, float freq, int polarity)
         close(enable_fd);
         close(period_fd);
         close(duty_fd);
+        char err[256];
+        snprintf(err, sizeof(err), "pwm_start: could not open polarity file");
+        add_error_msg(err);
         return -1;
     }
 
     // add to list
     new_pwm = malloc(sizeof(struct pwm_exp));
     if (new_pwm == 0) {
-        return -1; // out of memory
+        char err[256];
+        snprintf(err, sizeof(err), "pwm_start: unable to allocate memory");
+        add_error_msg(err);
+        return -1;
     }
 
     if (DEBUG)
-        printf(" ** IN pwm_start: IF WE MAKE IT HERE, THE FILES WERE SUCCESSFULLY OPEN **\n");
+        printf(" ** pwm_start: sysfs files opened successfully **\n");
     strncpy(new_pwm->key, key, KEYLEN);  /* can leave string unterminated */
     new_pwm->key[KEYLEN] = '\0'; /* terminate string */
     new_pwm->period_fd = period_fd;
@@ -393,13 +535,17 @@ int pwm_start(const char *key, float duty, float freq, int polarity)
 
     int rtnval = 0;
     rtnval = pwm_set_enable(key, ENABLE);
-    rtnval = 0;
-    rtnval = pwm_set_frequency(key, freq);
-    rtnval = 0;
-    //rtnval = pwm_set_polarity(key, polarity);
-    //rtnval = 0;
-    rtnval = pwm_set_duty_cycle(key, duty);
-
+    // Fix for issue #53
+    if (rtnval != -1) {
+        rtnval = 0;
+        rtnval = pwm_set_frequency(key, freq);
+        if (rtnval != -1) {
+            rtnval = 0;
+            //rtnval = pwm_set_polarity(key, polarity);
+            //rtnval = 0;
+            rtnval = pwm_set_duty_cycle(key, duty);
+        }
+    }
     return rtnval;
 }
 
@@ -407,7 +553,7 @@ int pwm_disable(const char *key)
 {
     struct pwm_exp *pwm, *temp, *prev_pwm = NULL;
 
-    int fd, len;
+    int fd, len, e_no;
     char str_gpio[80];
     // Per https://github.com/NextThingCo/CHIP-linux/pull/4
     // we need to export 0 here to enable pwm0
@@ -421,18 +567,30 @@ int pwm_disable(const char *key)
 
     if ((fd = open("/sys/class/pwm/pwmchip0/unexport", O_WRONLY)) < 0)
     {
+        char err[256];
+        snprintf(err, sizeof(err), "pwm_disable: could not open unexport file");
+        add_error_msg(err);
         return -1;
     }
     len = snprintf(str_gpio, sizeof(str_gpio), "%d", gpio); BUF2SMALL(str_gpio);
-    ssize_t s = write(fd, str_gpio, len);  ASSRT(s == len);
+    ssize_t s = write(fd, str_gpio, len); e_no = errno;
     close(fd);
-
+    if (s != len) {
+        char err[256];
+        snprintf(err, sizeof(err), "pwm_disable: could not unexport pwm (%s)", strerror(e_no));
+        add_error_msg(err);
+        return -1;
+    }
+    
     // remove from list
     pwm = exported_pwms;
     while (pwm != NULL)
     {
         if (strcmp(pwm->key, key) == 0)
         {
+            if (DEBUG) {
+                printf(" ** pwm_disable: freeing memory %s\n", key);
+            }
             //close the fd
             close(pwm->enable_fd);
             close(pwm->period_fd);
@@ -464,11 +622,3 @@ void pwm_cleanup(void)
     }
 }
 
-void pwm_toggle_debug(void)
-{
-    if (DEBUG) {
-        DEBUG = 0;
-    } else {
-        DEBUG = 1;
-    }
-}